requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
openpyxl==3.1.2
python-dotenv==1.0.0
